utils function tests
